added choices to model and custom_command

+++ core/models.py
@@ -57,4 +57,4 @@ class Sensor(models.Model):
 class SensorData(models.Model):
     sensor = models.ForeignKey('Sensor', on_delete=models.CASCADE)
     timestamp = models.DateTimeField('Read date', auto_now=True)
-    value = models.FloatField()
+    value = models.IntegerField()
